Refactored config files. See below for more details
This commit moves some config options from the training config to the model config.
The idea is that all configs related to the model itself belong in the model config file and all configs related to the dataset or training parameters belong in the training config.
This will make the model config file more portable for downstream tasks (e.g. NER/SQuAD).

Added to the model config:
- lowercase (bool): is the model cased or uncased (lowercase)
- next_sentence (bool): should the model contain the next sentence prediction head
- model_name (str): model type (e.g. bert-large-uncased)
- vocab_file (path): path to vocab file used for embedding layer and tokenizer